+++ libft/ft_striteri.c
@@ -6,16 +6,16 @@
 /*   By: kesawada <kesawada@student.42tokyo.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/28 11:19:48 by kesawada          #+#    #+#             */
-/*   Updated: 2023/09/28 11:19:52 by kesawada         ###   ########.fr       */
+/*   Updated: 2023/09/30 12:23:02 by kesawada         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_striteri(char *s, void (*f)(unsigned int, char*))
+void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	unsigned int	len;
-	unsigned int	i;
+	size_t	len;
+	size_t	i;
 
 	if (!s || !f)
 		return ;
